% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS.Click.R
\name{RS.Click}
\alias{RS.Click}
\title{Click}
\usage{
RS.Click(xpath)
}
\arguments{
\item{xpath}{position that you want to click}
}
\value{
a web
}
\description{
Click
}
\examples{
RS.Click(xpath)
}
