% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS.Get_attr.R
\name{RS.get_attr}
\alias{RS.get_attr}
\title{Get attr}
\usage{
RS.get_attr(xpath, attr)
}
\arguments{
\item{xpath}{xpath}

\item{attr}{attr}
}
\value{
string
}
\description{
Get attr
}
\examples{
RS.get_attr(xpath,attr)
}
