% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS.GetText.R
\name{RS.GetText}
\alias{RS.GetText}
\title{Get text}
\usage{
RS.GetText(xpath)
}
\arguments{
\item{xpath}{xpath}
}
\value{
string
}
\description{
Get text
}
\examples{
RS.GetText(xpath)
}
