% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS.goForward.R
\name{RS.go_Forward}
\alias{RS.go_Forward}
\title{go Forward}
\usage{
RS.go_Forward(n = 1)
}
\arguments{
\item{n}{go Forward times}
}
\value{
a web
}
\description{
go Forward
}
\examples{
RS.go_Forward(2)
}
