% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS.InputText.R
\name{RS.input_Text}
\alias{RS.input_Text}
\title{Input text and do not press enter}
\usage{
RS.input_Text(xpath, text)
}
\arguments{
\item{xpath}{xpath of location}

\item{text}{text that you want to input}
}
\value{
a web
}
\description{
Input text and do not press enter
}
\examples{
RS.input_Text(xpath,text)
}
