% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS.InputText_enter.R
\name{RS.InputText_enter}
\alias{RS.InputText_enter}
\title{Input text and press enter}
\usage{
RS.InputText_enter(xpath, text)
}
\arguments{
\item{xpath}{xpath of location}

\item{text}{text that you want to input}
}
\value{
a web
}
\description{
Input text and press enter
}
\examples{
RS.InputText_enter(xpath,text)
}
