% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS.OpenChrome.R
\name{RS.OpenChrome}
\alias{RS.OpenChrome}
\title{Open Chrome}
\usage{
RS.OpenChrome(port = 4444)
}
\arguments{
\item{port}{port, defaulted is 4444}
}
\value{
a web
}
\description{
Open Chrome
}
\examples{
RS.OpenChrome()
}
