% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS.OpenChrome_DisablePicture.R
\name{RS.OpenChrome_DisablePicture}
\alias{RS.OpenChrome_DisablePicture}
\title{Open Chrome without picture}
\usage{
RS.OpenChrome_DisablePicture(port = 4444)
}
\arguments{
\item{port}{port, defaulted is 4444}
}
\value{
a web
}
\description{
Open Chrome without picture
}
\examples{
RS.OpenChrome_DisablePicture()
}
