% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS.OpenFirefox.R
\name{RS.OpenFirefox}
\alias{RS.OpenFirefox}
\title{Open Chrome}
\usage{
RS.OpenFirefox(port = 4445)
}
\arguments{
\item{port}{port, defaulted is 4444}
}
\value{
a web
}
\description{
Open Chrome
}
\examples{
RS.OpenChrome()
}
