% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS.OpenUrl.R
\name{RS.OpenUrl}
\alias{RS.OpenUrl}
\title{Open url}
\usage{
RS.OpenUrl(url)
}
\arguments{
\item{url}{}
}
\value{
a web
}
\description{
Open url
}
\examples{
RS.OpenUrl('www.baidu.com')
}
