% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS.scroll.R
\name{RS.scroll_to_down_arrow}
\alias{RS.scroll_to_down_arrow}
\title{scroll down arrow}
\usage{
RS.scroll_to_down_arrow()
}
\value{
nothing
}
\description{
scroll down arrow
}
\examples{
scroll_to_down_arrow()
}
