% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS.scroll.R
\name{RS.scroll_to_home}
\alias{RS.scroll_to_home}
\title{scroll to the home}
\usage{
RS.scroll_to_home()
}
\value{
nothing
}
\description{
scroll to the home
}
\examples{
RS.scroll_to_home()
}
