% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS.scroll.R
\name{RS.scroll_to_up_arrow}
\alias{RS.scroll_to_up_arrow}
\title{scroll up arrow}
\usage{
RS.scroll_to_up_arrow()
}
\value{
nothing
}
\description{
scroll up arrow
}
\examples{
scroll_to_up_arrow（
}
